+++ source/databricks/tests/integration/calculator/test_calculator_job.py
@@ -27,14 +27,9 @@ from package.calculator_job import (
     _map_cim_quality_to_wholesale_quality,
 )
 from package.calculator_args import CalculatorArgs
-from package.codelists import TimeSeriesType, Grouping, TimeSeriesQuality
+from package.codelists import BasisDataType, TimeSeriesType, Grouping, TimeSeriesQuality
 import package.infrastructure as infra
 from package.schemas import time_series_point_schema, metering_point_period_schema
-from tests.helpers.file_utils import find_file
-from tests.helpers.assert_calculation_file_path import (
-    CalculationFileType,
-    assert_file_path_match_contract,
-)
 from typing import Callable, Optional
 from datetime import datetime
 
@@ -179,81 +174,95 @@ def test__get_valid_args_or_throw__accepts_parameters_from_process_manager(
     _get_valid_args_or_throw(dummy_job_parameters)
 
 
+@pytest.mark.parametrize(
+    "grid_area,energy_supplier_gln,balance_responsible_party_gln,time_series_type,grouping",
+    [
+        ("805", None, None, TimeSeriesType.PRODUCTION, Grouping.total_ga),
+        ("806", None, None, TimeSeriesType.PRODUCTION, Grouping.total_ga),
+        (
+            "805",
+            energy_supplier_gln_a,
+            None,
+            TimeSeriesType.NON_PROFILED_CONSUMPTION,
+            Grouping.es_per_ga,
+        ),
+        (
+            "806",
+            energy_supplier_gln_a,
+            None,
+            TimeSeriesType.NON_PROFILED_CONSUMPTION,
+            Grouping.es_per_ga,
+        ),
+        (
+            "805",
+            energy_supplier_gln_b,
+            None,
+            TimeSeriesType.NON_PROFILED_CONSUMPTION,
+            Grouping.es_per_ga,
+        ),
+        (
+            "806",
+            energy_supplier_gln_b,
+            None,
+            TimeSeriesType.NON_PROFILED_CONSUMPTION,
+            Grouping.es_per_ga,
+        ),
+        (
+            "805",
+            energy_supplier_gln_a,
+            balance_responsible_party_gln_a,
+            TimeSeriesType.PRODUCTION,
+            Grouping.es_per_brp_per_ga,
+        ),
+        (
+            "806",
+            energy_supplier_gln_a,
+            balance_responsible_party_gln_a,
+            TimeSeriesType.PRODUCTION,
+            Grouping.es_per_brp_per_ga,
+        ),
+        (
+            "805",
+            None,
+            None,
+            TimeSeriesType.NON_PROFILED_CONSUMPTION,
+            Grouping.total_ga,
+        ),
+        (
+            "806",
+            None,
+            None,
+            TimeSeriesType.NON_PROFILED_CONSUMPTION,
+            Grouping.total_ga,
+        ),
+    ],
+)
 def test__result_is_generated_for_requested_grid_areas(
     spark: SparkSession,
     data_lake_path: str,
     worker_id: str,
     executed_calculation_job: None,
+    grid_area: str,
+    energy_supplier_gln: str,
+    balance_responsible_party_gln: str,
+    time_series_type: TimeSeriesType,
+    grouping: str,
 ) -> None:
-    # Arrange
-    expected_ga_gln_type = [
-        ["805", None, None, TimeSeriesType.PRODUCTION, Grouping.total_ga],
-        ["806", None, None, TimeSeriesType.PRODUCTION, Grouping.total_ga],
-        [
-            "805",
-            energy_supplier_gln_a,
-            None,
-            TimeSeriesType.NON_PROFILED_CONSUMPTION,
-            Grouping.es_per_ga,
-        ],
-        [
-            "806",
-            energy_supplier_gln_a,
-            None,
-            TimeSeriesType.NON_PROFILED_CONSUMPTION,
-            Grouping.es_per_ga,
-        ],
-        [
-            "805",
-            energy_supplier_gln_b,
-            None,
-            TimeSeriesType.NON_PROFILED_CONSUMPTION,
-            Grouping.es_per_ga,
-        ],
-        [
-            "806",
-            energy_supplier_gln_b,
-            None,
-            TimeSeriesType.NON_PROFILED_CONSUMPTION,
-            Grouping.es_per_ga,
-        ],  # ga brp es
-        [
-            "805",
-            energy_supplier_gln_a,
-            balance_responsible_party_gln_a,
-            TimeSeriesType.PRODUCTION,
-            Grouping.es_per_brp_per_ga,
-        ],
-        [
-            "806",
-            energy_supplier_gln_a,
-            balance_responsible_party_gln_a,
-            TimeSeriesType.PRODUCTION,
-            Grouping.es_per_brp_per_ga,
-        ],
-    ]
-
     # Act
     # we run the calculator once per session. See the fixture executed_calculation_job in top of this file
 
     # Assert
-    for (
+    result_path = infra.get_result_file_relative_path(
+        executed_batch_id,
         grid_area,
         energy_supplier_gln,
         balance_responsible_party_gln,
         time_series_type,
         grouping,
-    ) in expected_ga_gln_type:
-        result_path = infra.get_result_file_relative_path(
-            executed_batch_id,
-            grid_area,
-            energy_supplier_gln,
-            balance_responsible_party_gln,
-            time_series_type,
-            grouping,
-        )
-        result = spark.read.json(f"{data_lake_path}/{worker_id}/{result_path}")
-        assert result.count() >= 1, "Calculator job failed to write files"
+    )
+    print(result_path)
+    result = spark.read.json(f"{data_lake_path}/{worker_id}/{result_path}")
+    assert result.count() >= 1, "Calculator job failed to write files"
 
 
 def test__published_time_series_points_contract_matches_schema_from_input_time_series_points(
@@ -443,8 +452,8 @@ def test__creates_hour_for_total_ga__with_expected_columns_names(
     executed_calculation_job: None,
 ) -> None:
     # Arrange
-    basis_data_relative_path = infra.get_time_series_hour_for_total_ga_relative_path(
-        executed_batch_id, "805"
+    basis_data_relative_path = infra.get_basis_data_path(
+        BasisDataType.TimeSeriesHour, executed_batch_id, "805"
     )
 
     # Act
@@ -469,8 +478,8 @@ def test__creates_hour_for_es_per_ga__with_expected_columns_names(
     executed_calculation_job: None,
 ) -> None:
     # Arrange
-    basis_data_relative_path = infra.get_time_series_hour_for_es_per_ga_relative_path(
-        executed_batch_id, "805", energy_supplier_gln_a
+    basis_data_relative_path = infra.get_basis_data_path(
+        BasisDataType.TimeSeriesHour, executed_batch_id, "805", energy_supplier_gln_a
     )
 
     # Act
@@ -495,8 +504,8 @@ def test__creates_quarter_for_total_ga__with_expected_columns_names(
     executed_calculation_job: None,
 ) -> None:
     # Arrange
-    relative_path = infra.get_time_series_quarter_for_total_ga_relative_path(
-        executed_batch_id, "805"
+    relative_path = infra.get_basis_data_path(
+        BasisDataType.TimeSeriesQuarter, executed_batch_id, "805"
     )
 
     # Act
@@ -522,8 +531,8 @@ def test__creates_quarter_for_es_per_ga__with_expected_columns_names(
     executed_calculation_job: None,
 ) -> None:
     # Arrange
-    relative_path = infra.get_time_series_quarter_for_es_per_ga_relative_path(
-        executed_batch_id, "805", energy_supplier_gln_a
+    relative_path = infra.get_basis_data_path(
+        BasisDataType.TimeSeriesQuarter, executed_batch_id, "805", energy_supplier_gln_a
     )
 
     # Act
@@ -549,15 +558,11 @@ def test__creates_quarter_for_total_ga__per_grid_area(
     executed_calculation_job: None,
 ) -> None:
     # Arrange
-    basis_data_relative_path_805 = (
-        infra.get_time_series_quarter_for_total_ga_relative_path(
-            executed_batch_id, "805"
-        )
+    basis_data_relative_path_805 = infra.get_basis_data_path(
+        BasisDataType.TimeSeriesQuarter, executed_batch_id, "805"
     )
-    basis_data_relative_path_806 = (
-        infra.get_time_series_quarter_for_total_ga_relative_path(
-            executed_batch_id, "806"
-        )
+    basis_data_relative_path_806 = infra.get_basis_data_path(
+        BasisDataType.TimeSeriesQuarter, executed_batch_id, "806"
     )
 
     # Act
@@ -588,15 +593,11 @@ def test__creates_quarter_for_es_per_ga__per_energy_supplier(
     executed_calculation_job: None,
 ) -> None:
     # Arrange
-    basis_data_relative_path_a = (
-        infra.get_time_series_quarter_for_es_per_ga_relative_path(
-            executed_batch_id, "805", energy_supplier_gln_a
-        )
+    basis_data_relative_path_a = infra.get_basis_data_path(
+        BasisDataType.TimeSeriesQuarter, executed_batch_id, "805", energy_supplier_gln_a
     )
-    basis_data_relative_path_b = (
-        infra.get_time_series_quarter_for_es_per_ga_relative_path(
-            executed_batch_id, "805", energy_supplier_gln_b
-        )
+    basis_data_relative_path_b = infra.get_basis_data_path(
+        BasisDataType.TimeSeriesQuarter, executed_batch_id, "805", energy_supplier_gln_b
     )
 
     # Act
@@ -627,11 +628,11 @@ def test__creates_hour_for_total_ga__per_grid_area(
     executed_calculation_job: None,
 ) -> None:
     # Arrange
-    basis_data_relative_path_805 = (
-        infra.get_time_series_hour_for_total_ga_relative_path(executed_batch_id, "805")
+    basis_data_relative_path_805 = infra.get_basis_data_path(
+        BasisDataType.TimeSeriesHour, executed_batch_id, "805"
     )
     basis_data_relative_path_806 = (
-        infra.get_time_series_hour_for_total_ga_relative_path(executed_batch_id, "806")
+        infra.get_basis_data_path(BasisDataType.TimeSeriesHour, executed_batch_id, "806")
     )
 
     # Act
@@ -662,11 +663,11 @@ def test__creates_hour_for_es_per_ga__per_energy_supplier(
     executed_calculation_job: None,
 ) -> None:
     # Arrange
-    basis_data_relative_path_a = infra.get_time_series_hour_for_es_per_ga_relative_path(
-        executed_batch_id, "805", energy_supplier_gln_a
+    basis_data_relative_path_a = infra.get_basis_data_path(
+        BasisDataType.TimeSeriesHour, executed_batch_id, "805", energy_supplier_gln_a
     )
-    basis_data_relative_path_b = infra.get_time_series_hour_for_es_per_ga_relative_path(
-        executed_batch_id, "805", energy_supplier_gln_b
+    basis_data_relative_path_b = infra.get_basis_data_path(
+        BasisDataType.TimeSeriesHour, executed_batch_id, "805", energy_supplier_gln_b
     )
 
     # Act
@@ -697,8 +698,8 @@ def test__master_basis_data_for_total_ga_has_expected_columns_names(
     executed_calculation_job: None,
 ) -> None:
     # Arrange
-    basis_data_path = infra.get_master_basis_data_for_total_ga_relative_path(
-        executed_batch_id, "805"
+    basis_data_path = infra.get_basis_data_path(
+        BasisDataType.MasterBasisData, executed_batch_id, "805"
     )
 
     # Act
@@ -729,8 +730,8 @@ def test__master_basis_data_for_es_per_ga_has_expected_columns_names(
     executed_calculation_job: None,
 ) -> None:
     # Arrange
-    basis_data_path = infra.get_master_basis_data_for_es_per_ga_relative_path(
-        executed_batch_id, "805", energy_supplier_gln_a
+    basis_data_path = infra.get_basis_data_path(
+        BasisDataType.MasterBasisData, executed_batch_id, "805", energy_supplier_gln_a
     )
 
     # Act
@@ -760,11 +761,11 @@ def test__creates_master_basis_data_per_grid_area(
     executed_calculation_job: None,
 ) -> None:
     # Arrange
-    basis_data_path_805 = infra.get_master_basis_data_for_total_ga_relative_path(
-        executed_batch_id, "805"
+    basis_data_path_805 = infra.get_basis_data_path(
+        BasisDataType.MasterBasisData, executed_batch_id, "805"
     )
-    basis_data_path_806 = infra.get_master_basis_data_for_total_ga_relative_path(
-        executed_batch_id, "806"
+    basis_data_path_806 = infra.get_basis_data_path(
+        BasisDataType.MasterBasisData, executed_batch_id, "806"
     )
 
     # Act: Executed in fixture executed_calculation_job
@@ -785,146 +786,6 @@ def test__creates_master_basis_data_per_grid_area(
     assert (
         master_basis_data_806.count() >= 1
     ), "Calculator job failed to write master basis data files for grid area 806"
-
-
-def test__master_basis_data_for_total_ga_filepath_matches_contract(
-    data_lake_path: str,
-    worker_id: str,
-    contracts_path: str,
-    executed_calculation_job: None,
-) -> None:
-    # Arrange
-    master_basis_data_path = infra.get_master_basis_data_for_total_ga_relative_path(
-        executed_batch_id, "805"
-    )
-
-    # Act: Executed in fixture executed_calculation_job
-
-    # Assert
-    actual_file_path = find_file(
-        f"{data_lake_path}/{worker_id}/",
-        f"{master_basis_data_path}/part-*.csv",
-    )
-    assert_file_path_match_contract(
-        contracts_path, actual_file_path, CalculationFileType.MasterBasisDataForTotalGa
-    )
-
-
-def test__hourly_basis_data_for_total_ga_filepath_matches_contract(
-    data_lake_path: str,
-    worker_id: str,
-    contracts_path: str,
-    executed_calculation_job: None,
-) -> None:
-    # Arrange
-    relative_output_path = infra.get_time_series_hour_for_total_ga_relative_path(
-        executed_batch_id, "805"
-    )
-
-    # Act: Executed in fixture executed_calculation_job
-
-    # Assert
-    actual_file_path = find_file(
-        f"{data_lake_path}/{worker_id}", f"{relative_output_path}/part-*.csv"
-    )
-    assert_file_path_match_contract(
-        contracts_path, actual_file_path, CalculationFileType.TimeSeriesHourBasisData
-    )
-
-
-def test__quarterly_basis_data_for_total_ga_filepath_matches_contract(
-    data_lake_path: str,
-    worker_id: str,
-    contracts_path: str,
-    executed_calculation_job: None,
-) -> None:
-    # Arrange
-    relative_output_path = infra.get_time_series_quarter_for_total_ga_relative_path(
-        executed_batch_id, "805"
-    )
-
-    # Act: Executed in fixture executed_calculation_job
-
-    # Assert
-    actual_file_path = find_file(
-        f"{data_lake_path}/{worker_id}", f"{relative_output_path}/part-*.csv"
-    )
-    assert_file_path_match_contract(
-        contracts_path,
-        actual_file_path,
-        CalculationFileType.TimeSeriesQuarterBasisDataForTotalGa,
-    )
-
-
-def test__master_basis_data_for_es_per_ga_filepath_matches_contract(
-    data_lake_path: str,
-    worker_id: str,
-    contracts_path: str,
-    executed_calculation_job: None,
-) -> None:
-    # Arrange
-    master_basis_data_path = infra.get_master_basis_data_for_es_per_ga_relative_path(
-        executed_batch_id, "805", energy_supplier_gln_a
-    )
-
-    # Act: Executed in fixture executed_calculation_job
-
-    # Assert
-    actual_file_path = find_file(
-        f"{data_lake_path}/{worker_id}/",
-        f"{master_basis_data_path}/part-*.csv",
-    )
-    assert_file_path_match_contract(
-        contracts_path, actual_file_path, CalculationFileType.MasterBasisDataForEsPerGa
-    )
-
-
-def test__hourly_basis_data_for_es_per_ga_filepath_matches_contract(
-    data_lake_path: str,
-    worker_id: str,
-    contracts_path: str,
-    executed_calculation_job: None,
-) -> None:
-    # Arrange
-    relative_output_path = infra.get_time_series_hour_for_es_per_ga_relative_path(
-        executed_batch_id, "805", energy_supplier_gln_a
-    )
-
-    # Act: Executed in fixture executed_calculation_job
-
-    # Assert
-    actual_file_path = find_file(
-        f"{data_lake_path}/{worker_id}", f"{relative_output_path}/part-*.csv"
-    )
-    assert_file_path_match_contract(
-        contracts_path,
-        actual_file_path,
-        CalculationFileType.TimeSeriesHourBasisDataForEsPerGa,
-    )
-
-
-def test__quarterly_basis_data_for_es_per_ga_filepath_matches_contract(
-    data_lake_path: str,
-    worker_id: str,
-    contracts_path: str,
-    executed_calculation_job: None,
-) -> None:
-    # Arrange
-    relative_output_path = infra.get_time_series_quarter_for_es_per_ga_relative_path(
-        executed_batch_id, "805", energy_supplier_gln_a
-    )
-
-    # Act: Executed in fixture executed_calculation_job
-
-    # Assert
-    actual_file_path = find_file(
-        f"{data_lake_path}/{worker_id}", f"{relative_output_path}/part-*.csv"
-    )
-    assert_file_path_match_contract(
-        contracts_path,
-        actual_file_path,
-        CalculationFileType.TimeSeriesQuarterBasisDataForEsPerGa,
-    )
 
 
 @patch("package.calculator_job._get_valid_args_or_throw")