+++ source/dotnet/Services/Domain/ProcessStepResultAggregate/QuantityQuality.cs
@@ -11,9 +11,14 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-namespace Energinet.DataHub.Wholesale.Contracts;
 
-/// <summary>
-/// An immutable batch.
-/// </summary>
-public sealed record BatchDtoV2(Guid BatchNumber, DateTimeOffset PeriodStart, DateTimeOffset PeriodEnd, DateTimeOffset? ExecutionTimeStart, DateTimeOffset? ExecutionTimeEnd, BatchState ExecutionState, bool IsBasisDataDownloadAvailable, string[] GridAreaCodes, ProcessType ProcessType);
+namespace Energinet.DataHub.Wholesale.Domain.ProcessStepResultAggregate;
+
+public enum QuantityQuality
+{
+    Missing,
+    Estimated,
+    Measured,
+    Incomplete,
+    Calculated,
+}