+++ source/dotnet/Packages/Contracts.Documentation/release-notes/release-notes.md
@@ -1,8 +1,20 @@
 # Wholesale Contracts Release notes
 
-## Version 5.0.14
+## Version 2.1.2
 
-Added process type to contract.
+No functional changes.
+
+## Version 2.1.1
+
+A few fixes to the contract `CalculationResultCompleted`.
+
+## Version 2.1.0
+
+Add integration event contract `CalculationResultCompleted`.
+
+## Version 2.0.14
+
+No functional changes.
 
 ## Version 2.0.13